refactor: import and export types as needed
