+++ src/publishToAllTopicsOnly/mod.ts
@@ -1,4 +1,5 @@
 import type { Temporal } from "https://cdn.skypack.dev/@js-temporal/polyfill?dts"
+import type { PubSubEvent } from "../types.ts"
 import { PUBSUB_ALL_TOPICS } from "../constants.ts"
 import { publish } from "../index.ts"
