+++ src/subscribe/types.ts
@@ -1,4 +1,4 @@
-declare namespace Intl {
+export declare namespace Intl {
 	type ListType = "conjunction" | "disjunction" | "unit"
 
 	interface ListFormatOptions {
@@ -21,31 +21,4 @@ declare namespace Intl {
 			options?: ListFormatOptions,
 		): string[]
 	}
-}
-
-// There's gotta be a better way.
-declare namespace Temporal {
-	type ZonedDateTime = unknown
-}
-
-type PubSubEvent = {
-	id?: string
-	eventName: string
-	timestamp?: Temporal.ZonedDateTime
-	data?: {
-		[key: string]: unknown
-	}
-}
-
-type Subscriptions = {
-	[token: string]: (event: PubSubEvent) => void
-}
-
-type Topics = {
-	[topic: string]: Subscriptions
-}
-
-type Subscribers = {
-	once?: Topics
-	always?: Topics
 }
