+++ src/subscribeToAllTopics/mod.ts
@@ -1,5 +1,6 @@
 import { subscribe } from "../index.ts"
 import { PUBSUB_ALL_TOPICS } from "../constants.ts"
+import type {PubSubEvent} from "../types.ts"
 
 export default function subscribeToAllTopics(
 	token: string,
