+++ src/subscribers/mod.ts
@@ -1,6 +1,8 @@
-const subscribers = {
-	once: {} as Topics,
-	always: {} as Topics,
+import type {Subscribers} from "./types.ts"
+
+const subscribers: Subscribers = {
+	once: {},
+	always: {},
 }
 
 export default subscribers
