+++ src/subscribers/types.ts
@@ -1,0 +1,7 @@
+import type {Topics} from "../types.ts"
+
+export type Subscribers = {
+	once?: Topics
+	always?: Topics
+}
+
